requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
flask-mail>=0.9.1
openai-whisper>=20231117
torch>=2.1.0
torchaudio>=2.1.0
werkzeug>=3.0.1
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
psycopg2-binary>=2.9.9
itsdangerous>=2.1.2
email-validator>=2.1.0
pyannote.audio>=3.1.1
speechbrain>=0.5.16
huggingface-hub>=0.19.0
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0
pytest>=7.4.0
pytest-flask>=1.3.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0